Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <math.h>
 int main() {
-    double a;
-    double b;
+    int a;
+    int b;
     char character;
     scanf("%d %d %c",&a,&b,&character);
     if(character=='+'){
